pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "glm-ocr-api"
version = "1.0.0"
description = "Self-hosted FastAPI backend for zai-org/GLM-OCR"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.9",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "torch>=2.3.0",
    "torchvision>=0.18.0",
    "transformers>=4.46.0",
    "accelerate>=0.30.0",
    "safetensors>=0.4.0",
    "pillow>=10.0.0",
    "pdf2image>=1.17.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
